feat: displaying of school classes in schoolManagement template

+++ src/main/java/com/projetws/model/SchoolClassRepository.java
@@ -12,10 +12,10 @@ public interface SchoolClassRepository extends PagingAndSortingRepository<School
 //	public List<Country> findByCountryName(String countryName);
 //	public Country findByCountryId(String id);
 
-	public boolean existsBySchoolClassId(int i);
+	public boolean existsBySchoolClassId(long i);
 
 
-	public SchoolClass findBySchoolClassId(int id);
+	public SchoolClass findBySchoolClassId(long id);
 	public SchoolClass findBySchool(School school);
 
 	public List<SchoolClass> findAllBySchool(School school);
